requirements.txt
chromadb==0.5.4
openai==1.37.0
python-dotenv==1.0.1
Requests==2.32.3
volcengine_python_sdk==1.0.93
